chore: Break up codegen/index.ts.
